feat(mespapiers): Replace ScanResultWrapper by ScanResultDialog
